platforms: cortex: move shell headers to calibration
Shell functions are specific to calibration for now.
Thus move headers in calib_platform.h.
