+++ controllers/quadpid/calibration/calib_quadpid.c
@@ -10,6 +10,7 @@
 
 /* Project includes */
 #include "platform.h"
+#include "calibration/calib_platform.h"
 #include "calibration/calib_quadpid.h"
 
 /**
